# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YWC API",
        "description": "API documentation for YWC platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://console-ywc.nativeappdev.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/itinerary/routes/{route}/comments": {
            "get": {
                "tags": [
                    "Itinerary Comments"
                ],
                "summary": "Get comments for a route",
                "operationId": "1479a5bedb763536a7dba0a7fd5f2140",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itinerary Comments"
                ],
                "summary": "Create a comment on a route",
                "operationId": "e068e4b7a071664a78c1eb7cfb37b354",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "stop_id": {
                                        "description": "Optional: Stop ID to comment on",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "description": "Optional: Parent comment ID for replies",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Great route! Looking forward to trying it."
                                    },
                                    "attachments": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "url1",
                                            "url2"
                                        ]
                                    },
                                    "visibility": {
                                        "type": "string",
                                        "enum": [
                                            "crew",
                                            "public"
                                        ],
                                        "example": "public"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment posted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment posted."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itinerary/routes/{route}/comments/{comment}": {
            "put": {
                "tags": [
                    "Itinerary Comments"
                ],
                "summary": "Update a comment",
                "operationId": "0825dfdfe137ef8ec1c472198b3a66e5",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Updated comment text"
                                    },
                                    "visibility": {
                                        "type": "string",
                                        "enum": [
                                            "crew",
                                            "public"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "hidden",
                                            "flagged"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment updated."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Itinerary Comments"
                ],
                "summary": "Delete a comment",
                "operationId": "a2845f9c72e6d515c76bc0be2287c210",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment removed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment removed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}